pptx-rag/requirements.txt
faiss-cpu==1.13.2
langchain==1.2.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.5
langchain-mcp-adapters==0.2.1
langchain-openai==1.1.6
langchain-text-splitters==1.1.0
loguru==0.7.3
pydantic==2.9.2
pydantic-settings==2.12.0
pydantic_core==2.23.4
python-dotenv==1.2.1
python-magic==0.4.27
python-pptx==1.0.2
sentence-transformers==5.2.2
streamlit==1.53.1
Pillow==12.1.0
rank_bm25==0.2.2
openai
pymupdf
